July 2024 release

+++ incremental_parsing/generation/stack_completion_right_context.py
@@ -5,7 +5,7 @@ import json
 import random
 import traceback
 from pathlib import Path
-from typing import NamedTuple
+from typing import NamedTuple, Optional
 
 import datasets
 import transformers
@@ -15,7 +15,7 @@ from transformers import GenerationMixin, PreTrainedTokenizer, AutoTokenizer
 from incremental_parsing.evaluation.text_cuts import cut_text_random, cut_text_same_indentation_levels
 from incremental_parsing.generation.constrained_generation import do_prefix_suffix_constrained_generation
 from incremental_parsing.generation.utils import tokenizer_int64, create_balanced_context, \
-    try_incremental_unconstrained, get_p50_p90
+    try_incremental_unconstrained, get_p50_p90_mean_count
 from incremental_parsing.lex_earley.lark_grammar import get_python_context
 from incremental_parsing.lex_earley.lex_earley import LexEarleyAlgorithmContext, lex_earley_parse, get_and_reset_stats
 
@@ -35,6 +35,8 @@ class ProgramParameters(NamedTuple):
     device: str
     max_tokens: int
     nice_cuts: bool
+    num_token_finding_attempts: Optional[int]
+    redo_completed: bool
 
 
 class RunParameters(NamedTuple):
@@ -46,17 +48,21 @@ class RunParameters(NamedTuple):
     device: str
     max_tokens: int
     nice_cuts: bool
+    num_token_finding_attempts: Optional[int]
+    redo_completed: bool
 
 
 def run_stack_instance(run_parameters: RunParameters, dataset: Dataset,
                        context: LexEarleyAlgorithmContext,
-                       tokenizer: PreTrainedTokenizer, model: GenerationMixin):
+                       tokenizer: PreTrainedTokenizer, model: GenerationMixin,
+                       warmup: bool= False):
 
     save_path = run_parameters.results_path / str(run_parameters.data_index) / str(run_parameters.cut_index) \
                 / str(run_parameters.seed)
-    save_path.mkdir(parents=True, exist_ok=True)
+    if not warmup:
+        save_path.mkdir(parents=True, exist_ok=True)
 
-    if (save_path / "unconstrained").exists():
+    if not warmup and not run_parameters.redo_completed and (save_path / "unconstrained").exists():
         return  # Allow resuming when interrupted without re-doing work
 
     chosen_dataset_content = dataset[run_parameters.data_index]["content"]
@@ -66,14 +72,15 @@ def run_stack_instance(run_parameters: RunParameters, dataset: Dataset,
     if run_parameters.nice_cuts:
         prefix, middle, suffix = cut_text_same_indentation_levels(chosen_dataset_content + "\n")
     else:
-        prefix, middle, suffix = cut_text_random(chosen_dataset_content, 0, .9, .2)
+        prefix, middle, suffix = cut_text_random(chosen_dataset_content, 0, .9, .2, 100)
         suffix = suffix + "\n"
 
     transformers.set_seed(hash((run_parameters.data_index, run_parameters.cut_index, run_parameters.seed)) % (2 ** 32))
 
-    (save_path / "prefix").write_text(prefix)
-    (save_path / "ground_truth").write_text(middle)
-    (save_path / "suffix").write_text(suffix)
+    if not warmup:
+        (save_path / "prefix").write_text(prefix)
+        (save_path / "ground_truth").write_text(middle)
+        (save_path / "suffix").write_text(suffix)
 
     pre_input_ids, pre_attention_mask = tokenizer_int64(tokenizer, prefix)
     post_input_ids, post_attention_mask = tokenizer_int64(tokenizer, suffix)
@@ -92,16 +99,20 @@ def run_stack_instance(run_parameters: RunParameters, dataset: Dataset,
             prefix_text=prefix, suffix_text=suffix,
             pre_input_ids=pre_input_ids, post_input_ids=post_input_ids,
             beam_size=run_parameters.beam_size, max_generation_length=run_parameters.max_tokens,
-            debug=False
+            debug=False, num_token_finding_attempts=run_parameters.num_token_finding_attempts
         )
 
-    constrained_overhead_p50, constrained_overhead_p90 = get_p50_p90(constrained_overhead_per_tok_times)
-    constrained_overhead_eval_p50, constrained_overhead_eval_p90 = get_p50_p90(constrained_overhead_per_eval_times)
+    constrained_overhead_p50, constrained_overhead_p90, constrained_overhead_mean, constrained_overhead_num = get_p50_p90_mean_count(constrained_overhead_per_tok_times)
+    constrained_overhead_eval_p50, constrained_overhead_eval_p90, constrained_overhead_eval_mean, constrained_overhead_eval_num = get_p50_p90_mean_count(constrained_overhead_per_eval_times)
 
-    if new_output_text_constrained is not None:
-        (save_path / "constrained").write_text(new_output_text_constrained)
+    if not warmup:
+        if new_output_text_constrained is not None:
+            (save_path / "constrained").write_text(new_output_text_constrained)
 
-    (save_path / "full_constrained").write_text(full_constrained)
+        if full_constrained is not None:
+            (save_path / "full_constrained").write_text(full_constrained)
+        else:
+            (save_path / "constrained_max_num_attempts").touch()
 
     unconstrained_start = datetime.datetime.now()
 
@@ -121,26 +132,37 @@ def run_stack_instance(run_parameters: RunParameters, dataset: Dataset,
     only_correct_unconstrained, unconstrained_overhead_times = try_incremental_unconstrained(
         prefix, new_output_text_unconstrained, suffix)
 
-    unconstrained_overhead_p50, unconstrained_overhead_p90 = get_p50_p90(unconstrained_overhead_times)
+    unconstrained_overhead_p50, unconstrained_overhead_p90, unconstrained_overhead_mean, unconstrained_overhead_num = get_p50_p90_mean_count(unconstrained_overhead_times)
 
-    if only_correct_unconstrained is not None:
-        (save_path / "unconstrained_checked").write_text(only_correct_unconstrained)
+    if not warmup:
+        if only_correct_unconstrained is not None:
+            (save_path / "unconstrained_checked").write_text(only_correct_unconstrained)
 
-    (save_path / "stats.json").write_text(json.dumps({
-        "pre_time": pre_time.total_seconds(),
-        "total_constrained_time": full_constrained_time.total_seconds(),
-        "constrained_overhead_p50": constrained_overhead_p50,
-        "constrained_overhead_p90": constrained_overhead_p90,
-        "constrained_overhead_eval_p50": constrained_overhead_eval_p50,
-        "constrained_overhead_eval_p90": constrained_overhead_eval_p90,
-        "num_constrained_tokens": len(tokenizer.encode(full_constrained)),
-        "num_unconstrained_tokens": len(new_output_tokens_unconstrained),
-        "unconstrained_generation_time": (unconstrained_end - unconstrained_start).total_seconds(),
-        "unconstrained_checking_overhead_p50": unconstrained_overhead_p50,
-        "unconstrained_checking_overhead_p90": unconstrained_overhead_p90,
-        **get_and_reset_stats()}))
+        (save_path / "stats.json").write_text(json.dumps({
+            "pre_time": pre_time.total_seconds(),
+            "total_constrained_time": full_constrained_time.total_seconds(),
+            "constrained_overhead_p50": constrained_overhead_p50,
+            "constrained_overhead_p90": constrained_overhead_p90,
+            "constrained_overhead_mean": constrained_overhead_mean,
+            "constrained_overhead_num": constrained_overhead_num,  # Should equal num_constrained_tokens, but just making sure
+            "constrained_overhead_eval_p50": constrained_overhead_eval_p50,
+            "constrained_overhead_eval_p90": constrained_overhead_eval_p90,
+            "constrained_overhead_eval_mean": constrained_overhead_eval_mean,
+            "constrained_overhead_eval_num": constrained_overhead_eval_num,
+            "num_constrained_tokens": len(tokenizer.encode(full_constrained)) if full_constrained is not None else None,
+            "num_constrained_chars": len(full_constrained) if full_constrained is not None else None,
+            "num_constrained_output_chars": len(new_output_text_constrained) if new_output_text_constrained is not None else None,
+            "num_unconstrained_tokens": len(new_output_tokens_unconstrained),
+            "num_unconstrained_chars": len(new_output_text_unconstrained),
+            "num_unconstrained_output_chars": len(only_correct_unconstrained) if only_correct_unconstrained is not None else None,
+            "unconstrained_generation_time": (unconstrained_end - unconstrained_start).total_seconds(),
+            "unconstrained_checking_overhead_p50": unconstrained_overhead_p50,
+            "unconstrained_checking_overhead_p90": unconstrained_overhead_p90,
+            "unconstrained_checking_overhead_mean": unconstrained_overhead_mean,
+            "unconstrained_checking_overhead_num": unconstrained_overhead_num,  # Should equal num unconstrained characters
+            **get_and_reset_stats()}, indent=1))
 
-    (save_path / "unconstrained").write_text(new_output_text_unconstrained)
+        (save_path / "unconstrained").write_text(new_output_text_unconstrained)
 
 
 def run_stack(program_parameters: ProgramParameters):
@@ -153,11 +175,13 @@ def run_stack(program_parameters: ProgramParameters):
     dataset = datasets.load_dataset(program_parameters.dataset_name, data_dir=program_parameters.dataset_path)["train"]
     earley_context = get_python_context()
 
+    has_warmed_up = False
+
     for data_index in range(program_parameters.min_data_idx, program_parameters.max_data_idx + 1):
         pyparse_err_path = program_parameters.results_path / f"{data_index}.pyparse.err"
         earleyparse_err_path = program_parameters.results_path / f"{data_index}.earleyparse.err"
 
-        if pyparse_err_path.exists() or earleyparse_err_path.exists():
+        if not program_parameters.redo_completed and (pyparse_err_path.exists() or earleyparse_err_path.exists()):
             continue  # Already looked at this and it is invalid
 
         # If the file doesn't parse, then we can't expect our parser to work so skip it
@@ -185,9 +209,6 @@ def run_stack(program_parameters: ProgramParameters):
             for seed in range(program_parameters.min_seed, program_parameters.max_seed + 1):
                 run_err_path = program_parameters.results_path / f"{data_index}.{cut_idx}.{seed}.err"
 
-                if run_err_path.exists():
-                    continue
-
                 run_parameters = RunParameters(
                     seed=seed,
                     cut_index=cut_idx,
@@ -196,14 +217,26 @@ def run_stack(program_parameters: ProgramParameters):
                     results_path=program_parameters.results_path,
                     device=program_parameters.device,
                     max_tokens=program_parameters.max_tokens,
-                    nice_cuts=program_parameters.nice_cuts
+                    nice_cuts=program_parameters.nice_cuts,
+                    num_token_finding_attempts=program_parameters.num_token_finding_attempts,
+                    redo_completed=program_parameters.redo_completed
                 )
+
+                if not has_warmed_up:
+                    # Get the juices flowing!
+                    run_stack_instance(run_parameters=run_parameters, dataset=dataset, context=earley_context,
+                                       tokenizer=tokenizer, model=model, warmup=True)
+                    has_warmed_up = True
+
+                if not program_parameters.redo_completed and run_err_path.exists():
+                    continue
 
                 try:
                     run_stack_instance(run_parameters=run_parameters, dataset=dataset,
                                        context=earley_context,
                                        tokenizer=tokenizer, model=model)
                 except Exception as e:
+                    # Likely cause: terminating search too early for a token (all candidates rejected)
                     run_err_path.write_text(traceback.format_exc())
 
 
@@ -223,6 +256,8 @@ def parse_args() -> ProgramParameters:
     parser.add_argument("--device", type=str)
     parser.add_argument("--max-tokens", type=int)
     parser.add_argument("--nice-cuts", action='store_true')
+    parser.add_argument("--redo-completed", action='store_true')
+    parser.add_argument("--num-token-finding-attempts", type=int, default=None)
     args = parser.parse_args()
 
     return ProgramParameters(**vars(args))
